backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# AI/LLM Integration
google-generativeai>=0.3.2
langchain==0.0.335
tiktoken==0.5.2
sentence-transformers==2.2.2
scikit-learn==1.3.2
joblib==1.3.2
numpy==1.26.1

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP Requests
httpx==0.25.2
requests==2.31.0

# Utilities
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# CORS
fastapi-cors==0.0.6
